+++ test/datalevin/search_test.clj
@@ -9,6 +9,7 @@
    [clojure.data.csv :as csv]
    [clojure.java.io :as io]
    [clojure.string :as s]
+   [cheshire.core :as json]
    [datalevin.test.core :as tdc :refer [db-fixture]]
    [clojure.test.check.generators :as gen]
    [clojure.test.check.clojure-test :as test]
@@ -350,6 +351,36 @@
                   [?e :id ?i]
                   [?e :description ?d]]
                 (d/db conn) "42")
+           "This is a new description"))
+    (d/close conn)
+    (u/delete-files dir)))
+
+(deftest load-json-test
+  (let [dir   (u/tmp-dir (str "load-json-test-" (UUID/randomUUID)))
+        conn  (d/create-conn
+                dir {:id          {:db/valueType :db.type/long
+                                   :db/unique    :db.unique/identity}
+                     :description {:db/valueType :db.type/string
+                                   :db/fulltext  true}})
+        data  (json/parse-string (slurp "test/data/data.json") true)
+        start (System/currentTimeMillis)]
+    (d/transact! conn data)
+    (is (< (- (System/currentTimeMillis) start) 5000))
+    (is (= 1 (count (d/fulltext-datoms (d/db conn)
+                                       "GraphstatsR" {:top 1000}))))
+    (is (= (d/q '[:find (count ?e) .
+                  :in $ ?q
+                  :where [(fulltext $ ?q {:top 1000}) [[?e _ _]]]]
+                (d/db conn) "GraphstatsR")
+           1))
+    (d/transact! conn [{:id          6299
+                        :description "This is a new description"}])
+    (is (= (d/q '[:find ?d .
+                  :in $ ?i
+                  :where
+                  [?e :id ?i]
+                  [?e :description ?d]]
+                (d/db conn) 6299)
            "This is a new description"))
     (d/close conn)
     (u/delete-files dir)))