+++ src/datalevin/main.clj
@@ -27,7 +27,7 @@
   (require 'datalevin.binding.graal)
   (require 'datalevin.binding.java))
 
-(def ^:private version "0.8.2")
+(def ^:private version "0.8.3")
 
 (def ^:private version-str
   (str