+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## WIP
+## 0.8.3 (2023-01-20)
 ### Fixed
 - [Datalog] error when large `:db/fulltext` value is added then removed in the same transaction.